requirements.txt
pandas==2.2.2
numpy==2.1.1
scipy==1.13.1
scikit-learn==1.5.2
joblib==1.4.2
pyyaml==6.0.2
